Inline single-use data provider traits (#631)
