+++ tests/Application/AbstractCreateSourceTest.php
@@ -5,15 +5,16 @@ declare(strict_types=1);
 namespace App\Tests\Application;
 
 use App\Entity\SourceInterface;
+use App\Enum\Source\Type;
 use App\Repository\SourceRepository;
-use App\Tests\DataProvider\CreateSourceInvalidRequestDataProviderTrait;
-use App\Tests\DataProvider\CreateSourceSuccessDataProviderTrait;
+use App\Request\FileSourceRequest;
+use App\Request\GitSourceRequest;
+use App\Request\InvalidSourceTypeRequest;
+use App\Request\SourceRequestInterface;
+use App\Tests\Services\SourceUserIdMutator;
 
 abstract class AbstractCreateSourceTest extends AbstractApplicationTest
 {
-    use CreateSourceInvalidRequestDataProviderTrait;
-    use CreateSourceSuccessDataProviderTrait;
-
     /**
      * @dataProvider createSourceInvalidRequestDataProvider
      *
@@ -28,6 +29,47 @@ abstract class AbstractCreateSourceTest extends AbstractApplicationTest
         );
 
         $this->responseAsserter->assertInvalidRequestJsonResponse($response, $expectedResponseData);
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function createSourceInvalidRequestDataProvider(): array
+    {
+        return [
+            'invalid source type' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => 'invalid',
+                ],
+                'expectedResponseData' => [
+                    'error' => [
+                        'type' => 'invalid_request',
+                        'payload' => [
+                            'type' => [
+                                'value' => 'invalid',
+                                'message' => InvalidSourceTypeRequest::ERROR_MESSAGE,
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            'git source missing host url' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
+                ],
+                'expectedResponseData' => [
+                    'error' => [
+                        'type' => 'invalid_request',
+                        'payload' => [
+                            'host-url' => [
+                                'value' => '',
+                                'message' => 'This value should not be blank.',
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ];
     }
 
     /**
@@ -59,5 +101,59 @@ abstract class AbstractCreateSourceTest extends AbstractApplicationTest
         $expected['user_id'] = $this->authenticationConfiguration->authenticatedUserId;
 
         $this->responseAsserter->assertSuccessfulJsonResponse($response, $expected);
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function createSourceSuccessDataProvider(): array
+    {
+        $hostUrl = 'https://example.com/repository.git';
+        $path = '/';
+        $credentials = md5((string) rand());
+        $label = 'file source label';
+
+        return [
+            'git source, credentials missing' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
+                    GitSourceRequest::PARAMETER_HOST_URL => $hostUrl,
+                    GitSourceRequest::PARAMETER_PATH => $path
+                ],
+                'expected' => [
+                    'user_id' => SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER,
+                    'type' => Type::GIT->value,
+                    'host_url' => $hostUrl,
+                    'path' => $path,
+                    'has_credentials' => false,
+                ],
+            ],
+            'git source, credentials present' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
+                    GitSourceRequest::PARAMETER_HOST_URL => $hostUrl,
+                    GitSourceRequest::PARAMETER_PATH => $path,
+                    GitSourceRequest::PARAMETER_CREDENTIALS => $credentials,
+                ],
+                'expected' => [
+                    'user_id' => SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER,
+                    'type' => Type::GIT->value,
+                    'host_url' => $hostUrl,
+                    'path' => $path,
+                    'has_credentials' => true,
+                ],
+            ],
+            'file source' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::FILE->value,
+                    FileSourceRequest::PARAMETER_LABEL => $label
+                ],
+                'expected' => [
+                    'user_id' => SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER,
+                    'type' => Type::FILE->value,
+                    'label' => $label,
+                ],
+            ],
+        ];
     }
 }
