+++ tests/Application/AbstractGetSourceTest.php
@@ -4,15 +4,18 @@ declare(strict_types=1);
 
 namespace App\Tests\Application;
 
+use App\Entity\FileSource;
+use App\Entity\GitSource;
+use App\Entity\RunSource;
 use App\Entity\SourceInterface;
+use App\Enum\RunSource\FailureReason;
+use App\Enum\RunSource\State;
+use App\Enum\Source\Type;
 use App\Model\EntityId;
-use App\Tests\DataProvider\GetSourceSuccessDataProviderTrait;
 use App\Tests\Services\SourceUserIdMutator;
 
 abstract class AbstractGetSourceTest extends AbstractApplicationTest
 {
-    use GetSourceSuccessDataProviderTrait;
-
     public function testGetSourceNotFound(): void
     {
         $response = $this->applicationClient->makeGetSourceRequest(
@@ -43,5 +46,70 @@ abstract class AbstractGetSourceTest extends AbstractApplicationTest
         $expectedResponseData = $sourceUserIdMutator->setSourceDataUserId($expectedResponseData);
 
         $this->responseAsserter->assertSuccessfulJsonResponse($response, $expectedResponseData);
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function getSourceSuccessDataProvider(): array
+    {
+        $userId = SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER;
+
+        $gitSource = new GitSource($userId, 'https://example.com/repository.git', '/', md5((string) rand()));
+        $fileSource = new FileSource($userId, 'file source label');
+        $runSource = new RunSource($fileSource);
+
+        $failureMessage = 'fatal: repository \'http://example.com/repository.git\' not found';
+        $failedRunSource = (new RunSource($gitSource))->setPreparationFailed(
+            FailureReason::GIT_CLONE,
+            $failureMessage
+        );
+
+        return [
+            Type::GIT->value => [
+                'source' => $gitSource,
+                'expectedResponseData' => [
+                    'id' => $gitSource->getId(),
+                    'user_id' => $gitSource->getUserId(),
+                    'type' => Type::GIT->value,
+                    'host_url' => $gitSource->getHostUrl(),
+                    'path' => $gitSource->getPath(),
+                    'has_credentials' => true,
+                ],
+            ],
+            Type::FILE->value => [
+                'source' => $fileSource,
+                'expectedResponseData' => [
+                    'id' => $fileSource->getId(),
+                    'user_id' => $fileSource->getUserId(),
+                    'type' => Type::FILE->value,
+                    'label' => $fileSource->getLabel(),
+                ],
+            ],
+            Type::RUN->value => [
+                'source' => $runSource,
+                'expectedResponseData' => [
+                    'id' => $runSource->getId(),
+                    'user_id' => $userId,
+                    'type' => Type::RUN->value,
+                    'parent' => $runSource->getParent()?->getId(),
+                    'parameters' => [],
+                    'state' => State::REQUESTED->value,
+                ],
+            ],
+            Type::RUN->value . ': preparation failed' => [
+                'source' => $failedRunSource,
+                'expectedResponseData' => [
+                    'id' => $failedRunSource->getId(),
+                    'user_id' => $userId,
+                    'type' => Type::RUN->value,
+                    'parent' => $failedRunSource->getParent()?->getId(),
+                    'parameters' => [],
+                    'state' => State::FAILED->value,
+                    'failure_reason' => FailureReason::GIT_CLONE->value,
+                    'failure_message' => $failureMessage,
+                ],
+            ],
+        ];
     }
 }
