--- tests/DataProvider/AddFileInvalidRequestDataProviderTrait.php
@@ -1,71 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\DataProvider;
-
-use App\Tests\Services\InvalidFilenameResponseDataFactory;
-use App\Validator\YamlFilenameConstraint;
-
-trait AddFileInvalidRequestDataProviderTrait
-{
-    /**
-     * @return array<mixed>
-     */
-    public function addFileInvalidRequestDataProvider(): array
-    {
-        return [
-            'name empty with .yaml extension, content non-empty' => [
-                'filename' => '.yaml',
-                'content' => 'non-empty value',
-                'expectedResponseData' => InvalidFilenameResponseDataFactory::createForMessage(
-                    YamlFilenameConstraint::MESSAGE_NAME_EMPTY
-                ),
-            ],
-            'name contains backslash characters, content non-empty' => [
-                'filename' => 'one-two-\\-three.yaml',
-                'content' => 'non-empty value',
-                'expectedResponseData' => InvalidFilenameResponseDataFactory::createForMessage(
-                    YamlFilenameConstraint::MESSAGE_FILENAME_INVALID
-                ),
-            ],
-            'name contains space characters, content non-empty' => [
-                'filename' => 'one two three.yaml',
-                'content' => 'non-empty value',
-                'expectedResponseData' => InvalidFilenameResponseDataFactory::createForMessage(
-                    YamlFilenameConstraint::MESSAGE_FILENAME_INVALID
-                ),
-            ],
-            'name valid, content empty' => [
-                'filename' => TestConstants::FILENAME,
-                'content' => '',
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'content' => [
-                                'value' => '',
-                                'message' => 'File content must not be empty.',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            'name valid, content invalid yaml' => [
-                'filename' => TestConstants::FILENAME,
-                'content' => "- item\ncontent",
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'content' => [
-                                'value' => '',
-                                'message' => 'Content must be valid YAML: Unable to parse at line 2 (near "content").',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-    }
-}
