--- tests/DataProvider/GetSourceSuccessDataProviderTrait.php
@@ -1,80 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\DataProvider;
-
-use App\Entity\FileSource;
-use App\Entity\GitSource;
-use App\Entity\RunSource;
-use App\Enum\RunSource\FailureReason;
-use App\Enum\RunSource\State;
-use App\Enum\Source\Type;
-
-trait GetSourceSuccessDataProviderTrait
-{
-    /**
-     * @return array<mixed>
-     */
-    public function getSourceSuccessDataProvider(): array
-    {
-        $userId = TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER;
-
-        $gitSource = new GitSource($userId, 'https://example.com/repository.git', '/', md5((string) rand()));
-        $fileSource = new FileSource($userId, 'file source label');
-        $runSource = new RunSource($fileSource);
-
-        $failureMessage = 'fatal: repository \'http://example.com/repository.git\' not found';
-        $failedRunSource = (new RunSource($gitSource))->setPreparationFailed(
-            FailureReason::GIT_CLONE,
-            $failureMessage
-        );
-
-        return [
-            Type::GIT->value => [
-                'source' => $gitSource,
-                'expectedResponseData' => [
-                    'id' => $gitSource->getId(),
-                    'user_id' => $gitSource->getUserId(),
-                    'type' => Type::GIT->value,
-                    'host_url' => $gitSource->getHostUrl(),
-                    'path' => $gitSource->getPath(),
-                    'has_credentials' => true,
-                ],
-            ],
-            Type::FILE->value => [
-                'source' => $fileSource,
-                'expectedResponseData' => [
-                    'id' => $fileSource->getId(),
-                    'user_id' => $fileSource->getUserId(),
-                    'type' => Type::FILE->value,
-                    'label' => $fileSource->getLabel(),
-                ],
-            ],
-            Type::RUN->value => [
-                'source' => $runSource,
-                'expectedResponseData' => [
-                    'id' => $runSource->getId(),
-                    'user_id' => $userId,
-                    'type' => Type::RUN->value,
-                    'parent' => $runSource->getParent()?->getId(),
-                    'parameters' => [],
-                    'state' => State::REQUESTED->value,
-                ],
-            ],
-            Type::RUN->value . ': preparation failed' => [
-                'source' => $failedRunSource,
-                'expectedResponseData' => [
-                    'id' => $failedRunSource->getId(),
-                    'user_id' => $userId,
-                    'type' => Type::RUN->value,
-                    'parent' => $failedRunSource->getParent()?->getId(),
-                    'parameters' => [],
-                    'state' => State::FAILED->value,
-                    'failure_reason' => FailureReason::GIT_CLONE->value,
-                    'failure_message' => $failureMessage,
-                ],
-            ],
-        ];
-    }
-}
