--- tests/DataProvider/TestConstants.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\DataProvider;
-
-class TestConstants
-{
-    public const FILENAME = 'filename.yaml';
-    public const AUTHENTICATED_USER_ID_PLACEHOLDER = '{{ authenticated_user_id }}';
-}
